split utils file into errorMEssages file and propertyValidationUtils
