+++ src/main/java/wolox/training/controllers/UserController.java
@@ -1,10 +1,11 @@
 package wolox.training.controllers;
 
+import static com.google.common.base.Strings.isNullOrEmpty;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -25,7 +26,6 @@ import wolox.training.models.Book;
 import wolox.training.models.User;
 import wolox.training.repositories.UserRepository;
 
-
 @RestController
 @RequestMapping("/api/users")
 @Api
@@ -44,7 +44,7 @@ public class UserController {
      */
     @GetMapping
     public Iterable findAll(@RequestParam(required = false) String username) {
-        if (username != null && !username.isEmpty())
+        if (!isNullOrEmpty(username))
             return userRepository.findByUsername(username);
         return userRepository.findAll();
     };
