+++ src/main/java/wolox/training/utils/PropertyValidationUtils.java
@@ -1,19 +1,13 @@
 package wolox.training.utils;
 
 import static com.google.common.base.Preconditions.checkArgument;
+import static wolox.training.utils.ErrorMessages.emptyErrorMessage;
+import static wolox.training.utils.ErrorMessages.nullErrorMessage;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDate;
 
-public class Utils {
-    public static String emptyErrorMessage(String identifier){
-        return String.format("%s cannot be empty!", identifier);
-    }
-
-    public static String nullErrorMessage(String identifier){
-        return String.format("%s cannot be null!", identifier);
-    }
-
+public class PropertyValidationUtils {
     public static LocalDate checkLocalDate(LocalDate date, String identifier){
         checkArgument(date != null, nullErrorMessage(identifier));
         return date;
